Paranoid make task sticky after scheduling
I don't think the task will be put back on the multiq once it has been
scheduled for the first time because the task's tid is set and I don't think
anything unsets it.

If the task has a tid set, then it effectively becomes sticky and stays on
that thread. This is useful for when @async is used in an Actor, as the async
task will be on the same thread as the Actor's main task.

However I have set sticky back to true to show intent, although this might
have some unintended consequences in the future.